Вынес из PrefKeys ProfileKeys
