+++ app/src/main/java/com/kawunus/glossario/ProfileKeys.kt
@@ -1,0 +1,14 @@
+package com.kawunus.glossario
+
+
+object ProfileKeys {
+    const val USER_EMAIL = "email"
+    const val USER_NICKNAME = "nickname"
+    const val USER_STATUS = "status"
+    const val USER_IMAGE = "profileImage"
+
+    object UserStatus {
+        const val REGISTER = "reg"
+        const val NOT_REGISTER = "noreg"
+    }
+}
